requirements.txt
fastapi
uvicorn
feedparser
newspaper3k
requests
